+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - InternalApi\ComponentType moved to Component\Component, and many static methods have been removed
 - Auth methods are no longer static
 - Auth login, signup, and logout methods now return a Symfony RedirectRepsonse object
+- component label is removed from metrics, add that in Prometheus instead
 
 ## 2.1.0 - 2025-03-23